code is split. working on refactor to mvp
