+++ components/connector/styles.module.css
@@ -1,0 +1,6 @@
+.connectors {
+  position: absolute;
+  height: 100vh;
+  width: 100vw;
+  pointer-events: none;
+}
