+++ components/connector/types.ts
@@ -1,0 +1,8 @@
+export interface ConnectorProps {
+  currentPath: {
+    x1: number;
+    y1: number;
+    x2: number;
+    y2: number;
+  };
+}
