+++ components/ports/styles.module.css
@@ -1,0 +1,20 @@
+.container {
+  position: absolute;
+  display: flex;
+  justify-content: space-evenly;
+  align-items: center;
+}
+
+.port {
+  height: var(--size-height-port);
+  width: var(--size-width-port);
+  border-style: solid;
+  border-radius: var(--size-border-radius-round);
+  border-width: var(--size-border-width-node);
+  border-color: var(--color-border-node);
+  background-color: var(--color-background-numberNode);
+}
+
+.port:hover {
+  cursor: crosshair;
+}
