+++ components/ports/types.ts
@@ -1,0 +1,6 @@
+export interface PortProps extends React.HTMLAttributes<HTMLButtonElement> {
+  amount: number;
+  horizontal?: boolean;
+  input?: boolean;
+  output?: boolean;
+}
